exception if invalid throttle parameter is used
If '*' is used as `throttle` parameter, there is a segfault
at the first request (see #350). This is a quick fix to avoid the
segfault exception.
Note that in general civetweb.c does not check parameters in advance,
in order to keep the code simple. Other invalid parameters may still
cause exceptions. It is the responsibility of the caller to pass
valid parameters to mg_start.