Timer test: Adapt to OSX clock_monotonic
It seems clock_monotonic value 123.456 is not in the past for OS-X systems,
as it is for Linux systems. Maybe it does not start on system startup, but
on process startup. Or maybe the OSX container environment used in
continuous integration testing are just a few seconds faster than Linux, and
this seems to make the difference here.
See https://travis-ci.org/civetweb/civetweb/builds/187029942