Allow "REPORT" HTTP method for REST calls
Some requets in a RESTful API should be a "GET", but they have a lot
of parameters, too long for the query string. The "GET" method does not
allow BODY with semantic meaning, so you cannot add these parameters as
JSON body.
A commonly used alternative would be to use "POST", but "POST" is neither
save nor idempotent, so it cannot be cached. Furthermore the expected
meaning of the words "POST" and "GET" is completely different.

Alternatively one can use the safe and idempotent method "REPORT",
defined in RFC3253, Section 3.6., allowing to cache responses.
"REPORT" has a similar meaning than "GET".

This commit allows scripts to handle REPORT methods.
Otherwise the server would filter them as "405 Method Not Allowed".

See also the list of all registered methods:
http://www.iana.org/assignments/http-methods/http-methods.xhtml

API calls that would be a GET in a RESTful API, but require a lot of parameter
cannot be made as GET, since the